server side fetch 리팩토링
