+++ src/app/Header.jsx
@@ -3,7 +3,7 @@ import HeaderLeft from '@/components/header/HeaderLeft';
 import HeaderCenter from '@/components/header/HeaderCenter';
 import HeaderRight from '@/components/header/HeaderRight';
 import MobileMenuList from '@/components/header/MobileMenuList';
-import { fetchSCSCGlobalStatus } from '@/util/fetchAPIData';
+import { fetchSCSCGlobalStatus } from '@/util/fetch/server-util';
 import styles from './Header.module.css';
 
 export default async function Header() {
