+++ src/app/api/executive/w/[name]/download/route.js
@@ -1,5 +1,6 @@
 import { getServerSession } from 'next-auth';
 import { authOptions } from '@/util/authOptions';
+import { fetchBackendServer } from '@/util/fetch/server';
 
 export async function GET(_, { params }) {
   const session = await getServerSession(authOptions);
@@ -8,18 +9,9 @@ export async function GET(_, { params }) {
     return Response.json({ detail: 'Unauthorized' }, { status: 401 });
   }
 
-  const hdrs = {};
-  hdrs['x-jwt'] = appJwt;
-
-  const res = await fetch(
-    `${process.env.BACKEND_URL || ''}/api/executive/w/${encodeURIComponent(
-      params['name'],
-    )}/download`,
-    {
-      method: 'GET',
-      headers: hdrs,
-      cache: 'no-store',
-    },
+  const res = await fetchBackendServer(
+    'GET',
+    `/api/executive/w/${encodeURIComponent(params['name'])}/download`,
   );
 
   const blob = await res.blob();
