+++ src/app/api/static/image/[...path]/route.js
@@ -1,23 +1,10 @@
-import { getServerSession } from 'next-auth';
-import { authOptions } from '@/util/authOptions';
+import { fetchBackendServer } from '@/util/fetch/server';
 
 export async function GET(_req, { params }) {
   const segments = Array.isArray(params?.path) ? params.path : [];
   const targetPath = segments.map(encodeURIComponent).join('/');
 
-  const base = process.env.BACKEND_URL || '';
-  const url = `${base}/static/image/${targetPath}`;
-
-  const session = await getServerSession(authOptions);
-  const appJwt = session?.backendJwt || null;
-
-  const res = await fetch(url, {
-    method: 'GET',
-    headers: {
-      ...(appJwt ? { 'x-jwt': appJwt } : {}),
-    },
-    cache: 'no-store',
-  });
+  const res = await fetchBackendServer('GET', `/static/image/${targetPath}`);
   return new Response(res.body, {
     status: res.status,
     headers: {
