+++ src/app/api/user/update-pfp-file/route.js
@@ -1,22 +1,13 @@
 export const runtime = 'nodejs';
 export const dynamic = 'force-dynamic';
 
-import { getServerSession } from 'next-auth';
-import { authOptions } from '@/util/authOptions';
+import { fetchBackendServer } from '@/util/fetch/server';
 
 export async function POST(request) {
-  const session = await getServerSession(authOptions);
-  const appJwt = session?.backendJwt || null;
-
   const formData = await request.formData();
 
-  const res = await fetch(`${process.env.BACKEND_URL || ''}/api/user/update-pfp-file`, {
-    method: 'POST',
-    headers: {
-      ...(appJwt ? { 'x-jwt': appJwt } : {}),
-    },
+  const res = await fetchBackendServer('POST', '/api/user/update-pfp-file', {
     body: formData,
-    cache: 'no-store',
   });
 
   return res;
