+++ src/app/board/[id]/page.jsx
@@ -1,18 +1,14 @@
 import './page.css';
 import BoardClient from '@/components/board/BoardClient';
+import { fetchBackendServer, fetchBackendServerJson } from '@/util/fetch/server';
 
 export async function generateMetadata({ params }) {
   const { id } = await params;
 
   try {
-    const res = await fetch(`${process.env.BACKEND_URL || ''}/api/board/${id}`, {
-      cache: 'no-store',
-    });
-    if (!res.ok) return { title: '게시판' };
-
-    const board = await res.json();
+    const board = await fetchBackendServerJson('GET', `/api/board/${id}`);
     const name = (board?.name || '').trim();
-    return { title: name || '게시판' }; // 루트 layout의 template이 있다면 자동으로 "이름 | SCSC"
+    return { title: name || '게시판' };
   } catch {
     return { title: '게시판' };
   }
@@ -22,12 +18,12 @@ export default async function BoardPage({ params }) {
   const { id } = await params;
   const boardId = id;
 
-  const boardRes = await fetch(`${process.env.BACKEND_URL || ''}/api/board/${boardId}`, {
-    cache: 'no-store',
-  });
+  const boardRes = await fetchBackendServer('GET', `/api/board/${boardId}`);
 
   if (!boardRes.ok) {
-    return <div className="text-center text-red-600 mt-10">게시글을 불러올 수 없습니다.</div>;
+    return (
+      <div className="text-center text-red-600 mt-10">게시판 정보를 불러올 수 없습니다.</div>
+    );
   }
 
   const board = await boardRes.json();
