+++ src/app/executive/page.jsx
@@ -2,7 +2,7 @@ import WithAuthorization from '@/components/WithAuthorization';
 import ScscStatusPanel from './ScscStatusPanel';
 import DiscordBotPanel from './DiscordBotPanel';
 import EnrollmentPolicyPanel from './EnrollmentPolicyPanel';
-import { fetchSCSCGlobalStatus, fetchDiscordBotStatus } from '@/util/fetchAPIData';
+import { fetchSCSCGlobalStatus, fetchDiscordBotStatus } from '@/util/fetch/server-util';
 import * as AdminLayout from '@/components/AdminLayout';
 
 const adminPageLinks = [
