+++ src/app/sig/[id]/page.jsx
@@ -1,19 +1,14 @@
 import 'highlight.js/styles/github.css';
 import './page.css';
 import SigClient from './SigClient';
-import { handleApiRequest } from '@/app/api/apiWrapper';
-import { fetchMe } from '@/util/fetchAPIData';
+import { fetchBackendServerJson } from '@/util/fetch/server';
+import { fetchCurrentUserProfile } from '@/util/fetch/server-util';
 import { redirect } from 'next/navigation';
 
 export async function generateMetadata({ params }) {
   const { id } = await params;
   try {
-    const res = await fetch(`${process.env.BACKEND_URL || ''}/api/sig/${id}`, {
-      method: 'GET',
-      cache: 'no-store',
-    });
-    if (!res.ok) throw new Error();
-    const sig = await res.json();
+    const sig = await fetchBackendServerJson('GET', `/api/sig/${id}`);
     return {
       title: sig.title,
       description: sig.description || 'SIG 상세 페이지',
@@ -49,15 +44,10 @@ export async function generateMetadata({ params }) {
 export default async function SigDetailPage({ params }) {
   const { id } = await params;
 
-  const [me] = await Promise.allSettled([fetchMe()]);
+  const me = await fetchCurrentUserProfile();
+  if (!me) redirect('/us/login');
 
-  if (me.status === 'rejected') redirect('/us/login');
-
-  const sigRes = await handleApiRequest('GET', `/api/sig/${id}`);
-  if (!sigRes.ok) {
-    return <div className="p-6 text-center text-red-600">존재하지 않는 SIG입니다.</div>;
-  }
-  const sig = await sigRes.json();
+  const sig = await fetchBackendServerJson('GET', `/api/sig/${id}`);
 
   const rawMembers = sig.members ?? [];
   const members = Array.isArray(rawMembers)
@@ -71,7 +61,7 @@ export default async function SigDetailPage({ params }) {
       sig={sig}
       members={members}
       articleContent={article.content}
-      me={me.value}
+      me={me}
       sigId={id}
     />
   );
